// src/models/Wildfire.ts
export class Wildfire {
  id: string | undefined;
  label: string | undefined;
  fireNumber: string | undefined;
  fireYear: number | undefined;
  fireStatus: string | undefined;
  areaEstimate: number | undefined;
  assessmentDate: string | undefined;
  generalCause: string | undefined;
  latitude: number | undefined;
  longitude: number | undefined;
  fireType: string | undefined;
  incidentType: number | undefined;
  responsibleArea: string | undefined;
}